+++ src/wine-host/utils.cpp
@@ -175,8 +175,8 @@ void MainContext::async_handle_watchdog_timer(
     // Try to keep a steady framerate, but add in delays to let other events
     // get handled if the GUI message handling somehow takes very long.
     watchdog_timer_.expires_at(std::chrono::steady_clock::now() + interval);
-    watchdog_timer_.async_wait([&](const boost::system::error_code& error) {
-        if (error.failed()) {
+    watchdog_timer_.async_wait([&](const std::error_code& error) {
+        if (error) {
             return;
         }